namespace and class names adjusted
