--- Protocol/data/rdapPort43.php
@@ -1,6 +1,0 @@
-<?php
-namespace Metaregistrar\RDAP {
-    class rdapPort43 extends rdapObject {
-
-    }
-}
