--- Protocol/rdapException.php
@@ -1,8 +1,0 @@
-<?php
-namespace Metaregistrar\RDAP {
-
-    class rdapException extends \Exception
-    {
-
-    }
-}
