--- Protocol/responses/rdapAsnResponse.php
@@ -1,6 +1,0 @@
-<?php
-namespace Metaregistrar\RDAP {
-    class rdapAsnResponse extends rdapResponse {
-
-    }
-}
