+++ composer.json
@@ -11,7 +11,10 @@
   ],
   "autoload": {
     "psr-4": {
-      "Metaregistrar\\": "Metaregistrar"
+      "Metaregistrar\\RDAP\\": "src"
     }
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^8"
   }
 }
