+++ src/Data/RdapPort43.php
@@ -1,0 +1,6 @@
+<?php
+namespace Metaregistrar\RDAP\Data;
+
+class RdapPort43 extends RdapObject
+{
+}
