+++ src/Data/RdapRole.php
@@ -1,0 +1,15 @@
+<?php
+namespace Metaregistrar\RDAP\Data;
+
+class RdapRole extends RdapObject
+{
+    public function dumpContents()
+    {
+        echo "- Role: ".$this->getRole()."\n";
+    }
+
+    public function getRole()
+    {
+        return $this->{0};
+    }
+}
