+++ src/RdapException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Metaregistrar\RDAP;
+
+class RdapException extends \Exception
+{
+}
