+++ src/Responses/RdapAsnResponse.php
@@ -1,0 +1,6 @@
+<?php
+namespace Metaregistrar\RDAP\Responses;
+
+class RdapAsnResponse extends RdapResponse
+{
+}
