+++ test/phpunit.xml
@@ -1,0 +1,18 @@
+<phpunit
+    bootstrap="Bootstrap.php"
+    colors="true"
+    convertErrorsToExceptions="true"
+    convertNoticesToExceptions="true"
+    convertWarningsToExceptions="true"
+>
+    <testsuites>
+        <testsuite name="RdapTest">
+            <directory>./</directory>
+        </testsuite>
+    </testsuites>
+    <filter>
+        <whitelist processUncoveredFilesFromWhitelist="false">
+            <directory suffix=".php">../</directory>
+        </whitelist>
+    </filter>
+</phpunit>
